First skeleton of CPU step with cycle counting

+++ nes/cpu.h
@@ -95,7 +95,7 @@ public:
 
     void Step();
 public:
-    u64 Cycles;
+    u32 Cycles;
 
 private:
     CpuRegs _regs;
